refactor: engine, separated managers, apis
Promote sideChannel to DataManager and eventChannel to EventManager, for
better reuse as needed.

Break down the entity manager into a component manager. Take the three
managers and wrap them up in an Engine class. The new Engine class
also has providers for DataManager, and EventManager. As well has
loopFixed for running a loop with a set number of iterations.

Putting these changes to use allows a great reduction in boilerplate
code for to actuall run the pipeline.